move wallets to their own dir
