+++ web/src/Msg.tsx
@@ -10,7 +10,7 @@ import {
     WormholeMessageId,
     chainToPlatform,
     signSendWait,
-    wormhole
+    wormhole,
 } from "@wormhole-foundation/sdk";
 import evm from "@wormhole-foundation/sdk/evm";
 import solana from "@wormhole-foundation/sdk/solana";
@@ -19,8 +19,8 @@ import { useEffect, useState } from "react";
 
 import "./App.css";
 import { NETWORK } from "./consts.js";
-import { MetaMaskSigner } from "./metamask.ts";
-import { PhantomProvider, PhantomSigner } from "./phantom.ts";
+import { MetaMaskSigner } from "./wallets/metamask.ts";
+import { PhantomProvider, PhantomSigner } from "./wallets/phantom.ts";
 
 const msk = new MetaMaskSDK();
 
@@ -35,19 +35,16 @@ function Msg() {
     useState<PhantomProvider | null>(null);
   const [solSigner, setSolSigner] = useState<SignAndSendSigner<
     typeof NETWORK,
-    PlatformToChains<"Solana"> 
+    PlatformToChains<"Solana">
   > | null>(null);
 
   // TODO: hardcoded for now, later allow selection of chains
-  const [chain] = useState<Chain>("Solana");
+  const [chain] = useState<"Solana">("Solana");
   // Set once the transfer is started
   const [srcTxIds, setSrcTxIds] = useState<string[]>([]);
   // Set once the transfer is attested
   const [attestations, setAttestations] = useState<WormholeMessageId[]>([]);
-  // Set after completing transfer
-  const [dstTxIds, setDstTxIds] = useState<string[]>([]);
-
-  const [wh, setWormhole] = useState<Wormhole<Network> | null>(null);
+  const [wh, setWormhole] = useState<Wormhole<typeof NETWORK> | null>(null);
 
   useEffect(() => {
     if (!wh) wormhole(NETWORK, [evm, solana]).then(setWormhole);
@@ -86,70 +83,58 @@ function Msg() {
     });
   }, [evmProvider, wh]);
 
-  function getSigner<C extends Chain, P extends ChainToPlatform<C>>(chain: C): Signer<typeof NETWORK, C> {
-    const platform: P = chainToPlatform(chain) as P
-
+  function getSigner<C extends Chain>(
+    chain: C
+  ): Signer<typeof NETWORK, C> {
     let s: Signer<typeof NETWORK, C> | null = null;
-    if(platform === "Evm"){
+    switch (chainToPlatform(chain)) {
+      case "Evm":
+        // TODO: actually check that the chain matches the one passed
         // @ts-ignore
-       s  = evmSigner;
-      (s as MetaMaskSigner<any>).requestChainChange(chain)
-    }else{
+        s = evmSigner;
+        (s as MetaMaskSigner<any>).requestChainChange(chain);
+        break;
+      case "Solana":
         // @ts-ignore
-        s = solSigner
+        s = solSigner;
+        break;
+      default:
+        throw new Error("No signer for: " + chain);
     }
-    if(s) return s;
 
-
-    throw new Error("No signer for: "+ chain)
+    return s!;
   }
-
 
   async function start(): Promise<void> {
     if (!wh) throw new Error("No wormhole");
 
-    const signer = getSigner(chain)
+    const signer = getSigner(chain);
+    console.log("Signer: ", signer);
     if (!signer) throw new Error("No signer");
 
-
-    // Create a transfer
     const chainCtx = wh.getChain(signer.chain());
     const snd = Wormhole.chainAddress(signer.chain(), signer.address());
     const core = await chainCtx.getWormholeCore();
-    // @ts-ignore
-    const xfer = core.publishMessage(snd, "Lol", 0, 0)
+
+    const xfer = core.publishMessage(snd.address, "Lol", 0, 0);
 
     // Start the transfer
     const txids = await signSendWait(chainCtx, xfer, signer);
     setSrcTxIds(txids.map((tx) => tx.txid));
 
     // Wait for attestation to be available
-    const att = await core.parseTransaction(txids[txids.length-1].txid);
+    const att = await core.parseTransaction(txids[txids.length - 1].txid);
     setAttestations(att as WormholeMessageId[]);
   }
 
-  //async function finish(): Promise<void> {
-  //  if (!wh) throw new Error("No wormhole");
-  //  if (!transfer) throw new Error("No Current transfer");
-
-  //  const signer = dstChain === "Solana" ? solSigner : evmSigner;
-  //  if (!signer) throw new Error("No signer");
-
-  //  // Finish transfer with updated signer
-  //  const finalTxs = await transfer.completeTransfer(signer);
-  //  setDstTxIds(finalTxs);
-  //}
-
   return (
     <>
-
       <div className="card">
-      <textarea></textarea>
+        <textarea></textarea>
         <button onClick={start} disabled={srcTxIds.length > 0}>
           Start transfer
         </button>
       </div>
-
     </>
   );
 }
